+++ src/metrics.rs
@@ -6,10 +6,10 @@ use std::sync::atomic::AtomicUsize;
 #[cfg(not(target_arch = "x86_64"))]
 use std::time::{Duration, Instant};
 
-#[cfg(feature = "no_metrics")]
+#[cfg(not(feature = "metrics"))]
 use std::marker::PhantomData;
 
-#[cfg(not(feature = "no_metrics"))]
+#[cfg(feature = "metrics")]
 use std::sync::atomic::Ordering::{Acquire, Relaxed};
 
 use crate::Lazy;
@@ -22,7 +22,7 @@ pub static M: Lazy<Metrics, fn() -> Metrics> = Lazy::new(Metrics::default);
 
 #[allow(clippy::cast_precision_loss)]
 pub(crate) fn clock() -> u64 {
-    if cfg!(feature = "no_metrics") {
+    if cfg!(not(feature = "metrics")) {
         0
     } else {
         #[cfg(target_arch = "x86_64")]
@@ -45,7 +45,7 @@ pub(crate) fn clock() -> u64 {
 pub(crate) fn uptime() -> Duration {
     static START: Lazy<Instant, fn() -> Instant> = Lazy::new(Instant::now);
 
-    if cfg!(feature = "no_metrics") {
+    if cfg!(not(feature = "metrics")) {
         Duration::new(0, 0)
     } else {
         START.elapsed()
@@ -55,9 +55,9 @@ pub(crate) fn uptime() -> Duration {
 /// Measure the duration of an event, and call `Histogram::measure()`.
 pub struct Measure<'h> {
     _start: u64,
-    #[cfg(not(feature = "no_metrics"))]
+    #[cfg(feature = "metrics")]
     histo: &'h Histogram,
-    #[cfg(feature = "no_metrics")]
+    #[cfg(not(feature = "metrics"))]
     _pd: PhantomData<&'h ()>,
 }
 
@@ -66,9 +66,9 @@ impl<'h> Measure<'h> {
     #[inline]
     pub fn new(_histo: &'h Histogram) -> Measure<'h> {
         Measure {
-            #[cfg(feature = "no_metrics")]
+            #[cfg(not(feature = "metrics"))]
             _pd: PhantomData,
-            #[cfg(not(feature = "no_metrics"))]
+            #[cfg(feature = "metrics")]
             histo: _histo,
             _start: clock(),
         }
@@ -78,7 +78,7 @@ impl<'h> Measure<'h> {
 impl<'h> Drop for Measure<'h> {
     #[inline]
     fn drop(&mut self) {
-        #[cfg(not(feature = "no_metrics"))]
+        #[cfg(feature = "metrics")]
         self.histo.measure(clock() - self._start);
     }
 }
@@ -141,7 +141,7 @@ pub struct Metrics {
     pub allocated_bytes: CachePadded<AtomicUsize>,
 }
 
-#[cfg(not(feature = "no_metrics"))]
+#[cfg(feature = "metrics")]
 impl Metrics {
     #[inline]
     pub fn tree_looped(&self) {
@@ -374,7 +374,7 @@ impl Metrics {
     }
 }
 
-#[cfg(feature = "no_metrics")]
+#[cfg(not(feature = "metrics"))]
 impl Metrics {
     pub const fn log_reservation_attempted(&self) {}